Refactor/richtexteditor (#505)
* Add rich text editor components and utilities

- Introduced `ContentRenderer` for rendering rich text with note links and markdown support.
- Added `NoteLinkCard` component for displaying linked notes.
- Created utility functions in `content-utils` for extracting prefixed IDs and converting between display and storage formats.
- Refactored `RichTextEditor` to utilize the new `ContentRenderer` and improved state management with `useRichTextEditor` hook.
- Updated types in `rich-text-editor.types.ts` for better type safety and clarity.

These changes enhance the functionality and usability of the rich text editor, allowing for better content linking and rendering capabilities.

* Refactor context search behavior in ChatContainer

- Updated the context search state to always enable by default, allowing users to control the feature independently of content context availability.
- Removed the auto-disable logic for context search when content context is present, simplifying user interaction and enhancing flexibility.

These changes improve the usability of the chat feature by providing users with more control over context search functionality.
